dont open console on startup
bug: cant log to it

+++ tanks/Transition.cpp
@@ -8,10 +8,10 @@ void Transition::render(sf::RenderTarget* g) {
 	sf::RectangleShape cover;
 	cover.setFillColor(sf::Color(30, 30, 30));
 	cover.setPosition(0, 0);
-	cover.setSize(sf::Vector2f(1280, y));
+	cover.setSize(sf::Vector2f(1280, (float)y));
 
 	sf::View viewZoomed = g->getView();
-	viewZoomed.zoom(1 - zoom);
+	viewZoomed.zoom((float)(1.0 - zoom));
 	g->setView(viewZoomed);
 	scene->render(g);
