+++ tanks/Transition.h
@@ -3,7 +3,7 @@
 class Transition : public Scene {
 private:
 	Scene* scene;
-	int y = 720;
+	double y = 720;
 	double zoom = 0.5;
 public:
 	Transition(Scene* scene);
